Rename to EmptyFreeformRuleset
