+++ templates/ruleset-empty/.template.config/template.json
@@ -11,6 +11,6 @@
         "language": "C#",
         "type": "project"
     },
-    "sourceName": "EmptyNonScrolling",
+    "sourceName": "EmptyFreeform",
     "preferNameDirectory": true
 }
