+++ templates/ruleset-empty/osu.Game.Rulesets.EmptyFreeformRuleset.Tests/TestSceneOsuGame.cs
@@ -1,0 +1,32 @@
+// Copyright (c) 2007-2019 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
+
+using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Shapes;
+using osu.Framework.Platform;
+using osu.Game.Tests.Visual;
+using osuTK.Graphics;
+
+namespace osu.Game.Rulesets.EmptyFreeformRuleset.Tests
+{
+    public class TestSceneOsuGame : OsuTestScene
+    {
+        [BackgroundDependencyLoader]
+        private void load(GameHost host, OsuGameBase gameBase)
+        {
+            OsuGame game = new OsuGame();
+            game.SetHost(host);
+
+            Children = new Drawable[]
+            {
+                new Box
+                {
+                    RelativeSizeAxes = Axes.Both,
+                    Colour = Color4.Black,
+                },
+                game
+            };
+        }
+    }
+}
