+++ templates/ruleset-empty/osu.Game.Rulesets.EmptyFreeformRuleset/EmptyFreeformRulesetRuleset.cs
@@ -1,0 +1,85 @@
+﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using System.Collections.Generic;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shapes;
+using osu.Framework.Graphics.Sprites;
+using osu.Framework.Input.Bindings;
+using osu.Game.Beatmaps;
+using osu.Game.Graphics;
+using osu.Game.Rulesets.Difficulty;
+using osu.Game.Rulesets.Mods;
+using osu.Game.Rulesets.EmptyFreeformRuleset.Beatmaps;
+using osu.Game.Rulesets.EmptyFreeformRuleset.Mods;
+using osu.Game.Rulesets.EmptyFreeformRuleset.UI;
+using osu.Game.Rulesets.UI;
+using osuTK;
+using osuTK.Graphics;
+
+namespace osu.Game.Rulesets.EmptyFreeformRuleset
+{
+    public class EmptyFreeformRulesetRuleset : Ruleset
+    {
+        public EmptyFreeformRulesetRuleset(RulesetInfo rulesetInfo = null)
+            : base(rulesetInfo)
+        {
+        }
+
+        public override string Description => "a very emptyfreeformruleset ruleset";
+
+        public override DrawableRuleset CreateDrawableRulesetWith(IWorkingBeatmap beatmap, IReadOnlyList<Mod> mods) =>
+            new DrawableEmptyFreeformRulesetRuleset(this, beatmap, mods);
+
+        public override IBeatmapConverter CreateBeatmapConverter(IBeatmap beatmap) =>
+            new EmptyFreeformRulesetBeatmapConverter(beatmap);
+
+        public override DifficultyCalculator CreateDifficultyCalculator(WorkingBeatmap beatmap) =>
+            new EmptyFreeformRulesetDifficultyCalculator(this, beatmap);
+
+        public override IEnumerable<Mod> GetModsFor(ModType type)
+        {
+            switch (type)
+            {
+                case ModType.Automation:
+                    return new[] { new EmptyFreeformRulesetModAutoplay() };
+
+                default:
+                    return new Mod[] { null };
+            }
+        }
+
+        public override string ShortName => "emptyfreeformruleset";
+
+        public override IEnumerable<KeyBinding> GetDefaultKeyBindings(int variant = 0) => new[]
+        {
+            new KeyBinding(InputKey.Z, EmptyFreeformRulesetAction.Button1),
+            new KeyBinding(InputKey.X, EmptyFreeformRulesetAction.Button2),
+        };
+
+        public override Drawable CreateIcon() => new Icon(ShortName[0]);
+
+        public class Icon : CompositeDrawable
+        {
+            public Icon(char c)
+            {
+                InternalChildren = new Drawable[]
+                {
+                    new Circle
+                    {
+                        Size = new Vector2(20),
+                        Colour = Color4.White,
+                    },
+                    new SpriteText
+                    {
+                        Anchor = Anchor.Centre,
+                        Origin = Anchor.Centre,
+                        Text = c.ToString(),
+                        Font = OsuFont.Default.With(size: 18)
+                    }
+                };
+            }
+        }
+    }
+}
