+++ templates/ruleset-empty/osu.Game.Rulesets.EmptyFreeformRuleset/UI/EmptyFreeformRulesetPlayfield.cs
@@ -1,0 +1,24 @@
+﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Game.Rulesets.UI;
+
+namespace osu.Game.Rulesets.EmptyFreeformRuleset.UI
+{
+    [Cached]
+    public class EmptyFreeformRulesetPlayfield : Playfield
+    {
+        protected override GameplayCursorContainer CreateCursor() => new EmptyFreeformRulesetCursorContainer();
+
+        [BackgroundDependencyLoader]
+        private void load()
+        {
+            AddRangeInternal(new Drawable[]
+            {
+                HitObjectContainer,
+            });
+        }
+    }
+}
