--- templates/ruleset-empty/osu.Game.Rulesets.EmptyNonScrolling.Tests/.vscode/launch.json
@@ -1,31 +1,0 @@
-{
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "name": "VisualTests (Debug)",
-            "type": "coreclr",
-            "request": "launch",
-            "program": "dotnet",
-            "args": [
-                "${workspaceRoot}/bin/Debug/netcoreapp2.2/osu.Game.Rulesets.EmptyNonScrolling.Tests.dll"
-            ],
-            "cwd": "${workspaceRoot}",
-            "preLaunchTask": "Build (Debug)",
-            "env": {},
-            "console": "internalConsole"
-        },
-        {
-            "name": "VisualTests (Release)",
-            "type": "coreclr",
-            "request": "launch",
-            "program": "dotnet",
-            "args": [
-                "${workspaceRoot}/bin/Release/netcoreapp2.2/osu.Game.Rulesets.EmptyNonScrolling.Tests.dll"
-            ],
-            "cwd": "${workspaceRoot}",
-            "preLaunchTask": "Build (Release)",
-            "env": {},
-            "console": "internalConsole"
-        }
-    ]
-}
