--- templates/ruleset-empty/osu.Game.Rulesets.EmptyNonScrolling.Tests/.vscode/tasks.json
@@ -1,47 +1,0 @@
-{
-    // See https://go.microsoft.com/fwlink/?LinkId=733558
-    // for the documentation about the tasks.json format
-    "version": "2.0.0",
-    "tasks": [
-        {
-            "label": "Build (Debug)",
-            "type": "shell",
-            "command": "dotnet",
-            "args": [
-                "build",
-                "--no-restore",
-                "osu.Game.Rulesets.EmptyNonScrolling.Tests.csproj",
-                "/p:GenerateFullPaths=true",
-                "/m",
-                "/verbosity:m"
-            ],
-            "group": "build",
-            "problemMatcher": "$msCompile"
-        },
-        {
-            "label": "Build (Release)",
-            "type": "shell",
-            "command": "dotnet",
-            "args": [
-                "build",
-                "--no-restore",
-                "osu.Game.Rulesets.EmptyNonScrolling.Tests.csproj",
-                "/p:Configuration=Release",
-                "/p:GenerateFullPaths=true",
-                "/m",
-                "/verbosity:m"
-            ],
-            "group": "build",
-            "problemMatcher": "$msCompile"
-        },
-        {
-            "label": "Restore",
-            "type": "shell",
-            "command": "dotnet",
-            "args": [
-                "restore"
-            ],
-            "problemMatcher": []
-        }
-    ]
-}
