--- templates/ruleset-empty/osu.Game.Rulesets.EmptyNonScrolling.Tests/VisualTestRunner.cs
@@ -1,23 +1,0 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using System;
-using osu.Framework;
-using osu.Framework.Platform;
-using osu.Game.Tests;
-
-namespace osu.Game.Rulesets.EmptyNonScrolling.Tests
-{
-    public static class VisualTestRunner
-    {
-        [STAThread]
-        public static int Main(string[] args)
-        {
-            using (DesktopGameHost host = Host.GetSuitableHost(@"osu", true))
-            {
-                host.Run(new OsuTestBrowser());
-                return 0;
-            }
-        }
-    }
-}
