--- templates/ruleset-empty/osu.Game.Rulesets.EmptyNonScrolling/EmptyNonScrollingDifficultyCalculator.cs
@@ -1,30 +1,0 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using System.Collections.Generic;
-using System.Linq;
-using osu.Game.Beatmaps;
-using osu.Game.Rulesets.Difficulty;
-using osu.Game.Rulesets.Difficulty.Preprocessing;
-using osu.Game.Rulesets.Difficulty.Skills;
-using osu.Game.Rulesets.Mods;
-
-namespace osu.Game.Rulesets.EmptyNonScrolling
-{
-    public class EmptyNonScrollingDifficultyCalculator : DifficultyCalculator
-    {
-        public EmptyNonScrollingDifficultyCalculator(Ruleset ruleset, WorkingBeatmap beatmap)
-            : base(ruleset, beatmap)
-        {
-        }
-
-        protected override DifficultyAttributes CreateDifficultyAttributes(IBeatmap beatmap, Mod[] mods, Skill[] skills, double clockRate)
-        {
-            return new DifficultyAttributes(mods, skills, 0);
-        }
-
-        protected override IEnumerable<DifficultyHitObject> CreateDifficultyHitObjects(IBeatmap beatmap, double clockRate) => Enumerable.Empty<DifficultyHitObject>();
-
-        protected override Skill[] CreateSkills(IBeatmap beatmap) => new Skill[0];
-    }
-}
