--- templates/ruleset-empty/osu.Game.Rulesets.EmptyNonScrolling/EmptyNonScrollingInputManager.cs
@@ -1,26 +1,0 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using System.ComponentModel;
-using osu.Framework.Input.Bindings;
-using osu.Game.Rulesets.UI;
-
-namespace osu.Game.Rulesets.EmptyNonScrolling
-{
-    public class EmptyNonScrollingInputManager : RulesetInputManager<EmptyNonScrollingAction>
-    {
-        public EmptyNonScrollingInputManager(RulesetInfo ruleset)
-            : base(ruleset, 0, SimultaneousBindingMode.Unique)
-        {
-        }
-    }
-
-    public enum EmptyNonScrollingAction
-    {
-        [Description("Button 1")]
-        Button1,
-
-        [Description("Button 2")]
-        Button2,
-    }
-}
