--- templates/ruleset-empty/osu.Game.Rulesets.EmptyNonScrolling/EmptyNonScrollingRuleset.cs
@@ -1,85 +1,0 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using System.Collections.Generic;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.Sprites;
-using osu.Framework.Input.Bindings;
-using osu.Game.Beatmaps;
-using osu.Game.Graphics;
-using osu.Game.Rulesets.Difficulty;
-using osu.Game.Rulesets.Mods;
-using osu.Game.Rulesets.EmptyNonScrolling.Beatmaps;
-using osu.Game.Rulesets.EmptyNonScrolling.Mods;
-using osu.Game.Rulesets.EmptyNonScrolling.UI;
-using osu.Game.Rulesets.UI;
-using osuTK;
-using osuTK.Graphics;
-
-namespace osu.Game.Rulesets.EmptyNonScrolling
-{
-    public class EmptyNonScrollingRuleset : Ruleset
-    {
-        public EmptyNonScrollingRuleset(RulesetInfo rulesetInfo = null)
-            : base(rulesetInfo)
-        {
-        }
-
-        public override string Description => "a very emptynonscrolling ruleset";
-
-        public override DrawableRuleset CreateDrawableRulesetWith(IWorkingBeatmap beatmap, IReadOnlyList<Mod> mods) =>
-            new DrawableEmptyNonScrollingRuleset(this, beatmap, mods);
-
-        public override IBeatmapConverter CreateBeatmapConverter(IBeatmap beatmap) =>
-            new EmptyNonScrollingBeatmapConverter(beatmap);
-
-        public override DifficultyCalculator CreateDifficultyCalculator(WorkingBeatmap beatmap) =>
-            new EmptyNonScrollingDifficultyCalculator(this, beatmap);
-
-        public override IEnumerable<Mod> GetModsFor(ModType type)
-        {
-            switch (type)
-            {
-                case ModType.Automation:
-                    return new[] {new EmptyNonScrollingModAutoplay()};
-
-                default:
-                    return new Mod[] {null};
-            }
-        }
-
-        public override string ShortName => "emptynonscrolling";
-
-        public override IEnumerable<KeyBinding> GetDefaultKeyBindings(int variant = 0) => new[]
-        {
-            new KeyBinding(InputKey.Z, EmptyNonScrollingAction.Button1),
-            new KeyBinding(InputKey.X, EmptyNonScrollingAction.Button2),
-        };
-
-        public override Drawable CreateIcon() => new Icon(ShortName[0]);
-
-        public class Icon : CompositeDrawable
-        {
-            public Icon(char c)
-            {
-                InternalChildren = new Drawable[]
-                {
-                    new Circle
-                    {
-                        Size = new Vector2(20),
-                        Colour = Color4.White,
-                    },
-                    new SpriteText
-                    {
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        Text = c.ToString(),
-                        Font = OsuFont.Default.With(size: 18)
-                    }
-                };
-            }
-        }
-    }
-}
